refactor(cli): extract db.py environment-diagnostic helpers to _db_diagnostics
Moves the four environment-diagnostic helpers + their probe table out
of amx/cli_support/commands/db.py into a focused
amx/cli_support/_db_diagnostics.py:

- _BACKEND_DRIVER_PROBES (the per-backend (apt-tag, importable-name)
  table)
- _print_system_prereq_hint (OS-level package hints for mssql /
  oracle / etc.)
- _offer_to_install_backend_driver (Python-side driver install button)
- _is_databricks_tls_failure (TLS-error pattern match)
- _env_trusted_ca_candidate (REQUESTS_CA_BUNDLE / SSL_CERT_FILE pickup)

db.py re-exports each name so existing imports and the test in
test_mssql_system_prereq_hint keep working. That test's
'amx.cli_support.commands.db.info' patch was rewritten to target
'amx.cli_support._db_diagnostics.info' since the info() call now
lives in the diagnostics module.

db.py: 1911 -> 1810 LOC (-101, -5%).
